the voids calculation gives packing indexes very close to Platon; some optimisation is implemented to cut off iterations over 'remote' atoms;

+++ xlib/unitcell.h
@@ -260,7 +260,7 @@ protected:
     }
   };
   class TBuildDistanceMapTask  {
-    short*** map;
+    float*** map;
     TTypeList<AnAssociation3<vec3f,TCAtom*, float> >& atoms;
     const vec3i& dims;
     const mat3f& tm;
@@ -270,7 +270,7 @@ protected:
     tm(parent.tm), map(parent.map), atoms(parent.atoms), dims(parent.dims),
       loop_data(parent.loop_data), owns_data(false)  {}
   public:
-    TBuildDistanceMapTask(const mat3f& _tm, short*** _map, const vec3i& _dims,
+    TBuildDistanceMapTask(const mat3f& _tm, float*** _map, const vec3i& _dims,
       TTypeList<AnAssociation3<vec3f,TCAtom*, float> >& _atoms) :
       tm(_tm), map(_map), atoms(_atoms), dims(_dims), owns_data(true)
     {
